moderator access

+++ api.js
@@ -7,6 +7,7 @@ var surveyController = require('./controllers/survey.js');
 var actionController = require('./controllers/actions.js');
 var tagController = require('./controllers/tags.js');
 var auth = require('./controllers/authorization.js');
+var accessController = require('./controllers/access.js');
 var config = require('./config.js');
 var confessionModel = require('./models/confession.js');
 var replyModel = require('./models/reply.js');
@@ -26,6 +27,7 @@ apiRouter.get('/participants/:entry_id', (req, res)=>{
 });
 apiRouter.use(auth);
 apiRouter.route('/confession/accept/:confession_id').get((req, res)=>{
+  if(!accessController(req.user.flags, 'addEntry'))return res.json({success: false, response: {message: 'You\'re not allowed to perform this action'}});
   confessionModel.findById(req.params.confession_id).populate('survey').exec((err, confession)=>{
     if(err) return res.send(err);
     if(confession.entryID && confession.status==1){
@@ -41,7 +43,7 @@ apiRouter.route('/confession/accept/:confession_id').get((req, res)=>{
       if(confession.embed){
         surveyController.uploadAttachment(confession.embed, function(attachment){
           if(attachment.success){
-            req.decoded._doc.embedHash = attachment.hash;
+            req.attachment = attachment;
             surveyController.acceptSurvey(confession, req, function(result){
               if(!result.success){
                 surveyController.wykopLogin();
@@ -69,12 +71,13 @@ apiRouter.route('/confession/accept/:confession_id').get((req, res)=>{
   })
 });
 apiRouter.route('/confession/danger/:confession_id').get((req, res)=>{
+  if(!accessController(req.user.flags, 'setStatus'))return res.json({success: false, response: {message: 'You\'re not allowed to perform this action'}});
   confessionModel.findById(req.params.confession_id, (err, confession)=>{
     if(err) return res.json(err);
     confession.status==-1?confession.status=0:confession.status=-1;
     var status = confession.status==0?'warning':'danger';
     var actionType = confession.status==0?3:2;
-    actionController(confession, req.decoded._doc._id, actionType);
+    actionController(confession, req.user._id, actionType);
     confession.save((err)=>{
       if(err) return res.json({success: false, response: {message: err}});
       res.json({success: true, response: {message: 'Zaaktualizowano status', status: status}});
@@ -82,10 +85,11 @@ apiRouter.route('/confession/danger/:confession_id').get((req, res)=>{
   });
 });
 apiRouter.route('/confession/tags/:confession_id/:tag').get((req, res)=>{
+  if(!accessController(req.user.flags, 'updateTags'))return res.json({success: false, response: {message: 'You\'re not allowed to perform this action'}});
   //there's probably more clean way to do this.
   confessionModel.findById(req.params.confession_id, (err, confession)=>{
     if(err)return res.send(err);
-    actionController(confession, req.decoded._doc._id, 9);
+    actionController(confession, req.user._id, 9);
     confessionModel.update({_id: req.params.confession_id}, {$set: {tags: tagController.prepareArray(confession.tags, req.params.tag)}}, (err)=>{
       if(err)return res.json({success: false, response: {message: err}});
       res.json({success: true, response: {message: 'Tagi zaaktualizowano', status: 'success'}});
@@ -93,21 +97,23 @@ apiRouter.route('/confession/tags/:confession_id/:tag').get((req, res)=>{
   });
 });
 apiRouter.route('/confession/delete/:confession_id').get((req, res)=>{
+  if(!accessController(req.user.flags, 'deleteEntry'))return res.json({success: false, response: {message: 'You\'re not allowed to perform this action'}});
   confessionModel.findById(req.params.confession_id, (err, confession)=>{
     if(err) return res.send(err);
     wykopController.deleteEntry(confession.entryID, (err, result, deletedEntry)=>{
       if(err) return res.json({success: false, response: {message: err.error.message}});
-        actionController(confession, req.decoded._doc._id, 5);
+        actionController(confession, req.user._id, 5);
         confession.status = -1;
         confession.save((err)=>{
           if(err)return res.json({success: false, response: {message: err}});
           res.json({success: true, response: {message: `Usunięto wpis ID: ${result.id}`}});
-          wykopController.sendPrivateMessage('sokytsinolop', `${req.decoded._doc.username} usunął wpis \n ${deletedEntry.id}`, ()=>{});
+          wykopController.sendPrivateMessage('sokytsinolop', `${req.user.username} usunął wpis \n ${deletedEntry.id}`, ()=>{});
       });
     });
   });
 });
 apiRouter.route('/reply/accept/:reply_id').get((req, res)=>{
+  if(!accessController(req.user.flags, 'addReply'))return res.json({success: false, response: {message: 'You\'re not allowed to perform this action'}});
   replyModel.findById(req.params.reply_id).populate('parentID').exec((err, reply)=>{
     if(err) return console.log(err);
     if(reply.commentID){
@@ -124,6 +130,7 @@ apiRouter.route('/reply/accept/:reply_id').get((req, res)=>{
   });
 });
 apiRouter.route('/reply/danger/:reply_id').get((req, res)=>{
+  if(!accessController(req.user.flags, 'setStatus'))return res.json({success: false, response: {message: 'You\'re not allowed to perform this action'}});
   replyModel.findById(req.params.reply_id).populate('parentID').exec((err, reply)=>{
     if(err) return console.log(err);
     var message = '';
@@ -132,7 +139,7 @@ apiRouter.route('/reply/danger/:reply_id').get((req, res)=>{
     var actionType = reply.status==0?3:2;
     reply.save((err)=>{
       if(err) res.json({success: false, response: {message: err}});
-      actionController(reply.parentID, req.decoded._doc._id, actionType);
+      actionController(reply.parentID, req.user._id, actionType);
       res.json({success: true, response: {message: 'Status zaaktualizowany', status: status}});
     });
   });
